Trait oak_proxy_attestation::proto::proxy_attestation_server::ProxyAttestation

pub trait ProxyAttestation: Send + Sync + 'static {
    fn get_signed_certificate<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetSignedCertificateRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetSignedCertificateResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_root_certificate<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetRootCertificateRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetRootCertificateResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ProxyAttestationServer.

Required methods

fn get_signed_certificate<'life0, 'async_trait>(
&'life0 self,
request: Request<GetSignedCertificateRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetSignedCertificateResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get signed X.509 certificate based on the provided certificate signing request.

fn get_root_certificate<'life0, 'async_trait>(
&'life0 self,
request: Request<GetRootCertificateRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetRootCertificateResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get root X.509 certificate used by the Proxy Attestation Service to sign certificates. This certificate is generated insode the TEE and is not the same certificate that is used to connect to the proxy itself.

Note: Current implementation doesn’t support sealing, so the proxy generates a new certificate on startup.